+++ SpatialGDK/Source/SpatialGDKTests/SpatialGDK/Interop/SpatialWorkerFlags/WorkerFlagsTestSpyObject.cpp
@@ -1,0 +1,15 @@
+// Copyright (c) Improbable Worlds Ltd, All Rights Reserved
+
+#include "WorkerFlagsTestSpyObject.h"
+
+void  UWorkerFlagsTestSpyObject::SetFlagUpdated(const FString& flagName, const FString& flagValue)
+{
+	timesUpdated++;
+
+	return;
+}
+
+int UWorkerFlagsTestSpyObject::getTimesFlagUpdated() const
+{
+	return timesUpdated;
+}